DEV: Extract `BuildWithAiModal` from workflows canvas
